refactor: move Imager to new imager.py
