+++ gmatpyplus/hardware/__init__.py
@@ -1,4 +1,5 @@
 from .fieldofview import *
 from .fueltank import *
 from .hardware import *
+from .imager import *
 from .thruster import *
